+++ components/messaging/ConversationProvider.tsx
@@ -1,39 +1,11 @@
 'use client';
-import React, { ReactNode, createContext, useEffect, useState } from 'react';
+import React, { ReactNode, useEffect, useState } from 'react';
 import {
   AllConversationsType,
   ConversationCardType,
 } from '@/utils/messaging/messagingTypes';
+import useConversation from './app/(dashboard)/conversations/useContext';
 import newClient from '@/config/supabaseclient';
-
-type ConversationProviderProps = {
-  allConversations: AllConversationsType;
-  setAllConversations: React.Dispatch<
-    React.SetStateAction<AllConversationsType>
-  >;
-  openConversation: ConversationCardType;
-  setOpenConversation: React.Dispatch<
-    React.SetStateAction<ConversationCardType>
-  > | null;
-};
-
-const defaultContext: ConversationProviderProps = {
-  allConversations: [],
-  setAllConversations: () => [],
-  openConversation: {
-    joined_at: new Date().toString(),
-    conversation_id: 2,
-    user_id: 'default',
-    conversations: {
-      id: 1,
-      messages: [],
-      created_at: new Date().toString(),
-    },
-  },
-  setOpenConversation: () => null,
-};
-
-export const ConversationContext = createContext(defaultContext);
 
 const ConversationProvider = ({
   children,
@@ -45,7 +17,16 @@ const ConversationProvider = ({
   const [allConversations, setAllConversations] =
     useState<AllConversationsType>([]);
   const [openConversation, setOpenConversation] =
-    useState<ConversationCardType>(null);
+    useState<ConversationCardType>({
+      joined_at: new Date().toString(),
+      conversation_id: 2,
+      user_id: 'default',
+      conversations: {
+        id: 1,
+        messages: [],
+        created_at: new Date().toString(),
+      },
+    });
 
   useEffect(() => {
     const fetchConversations = async () => {
@@ -57,7 +38,7 @@ const ConversationProvider = ({
           .eq('user_id', userId);
 
         setAllConversations(fetchedConversations ?? []);
-        setOpenConversation(allConversations[0]);
+        setOpenConversation && setOpenConversation(allConversations[0]);
       } catch (error) {
         console.error(`Failed to fetch conversations from database: ${error}`);
         throw error;
@@ -68,7 +49,7 @@ const ConversationProvider = ({
   }, []);
 
   return (
-    <ConversationContext.Provider
+    <useConversation.Provider
       value={{
         allConversations,
         setAllConversations,
@@ -77,7 +58,7 @@ const ConversationProvider = ({
       }}
     >
       {children}
-    </ConversationContext.Provider>
+    </useConversation.Provider>
   );
 };
