+++ components/messaging/ConversationsList.tsx
@@ -1,11 +1,10 @@
 'use client';
 import ConversationCard from './ConversationCard';
 import { useContext } from 'react';
-import { ConversationContext } from './ConversationProvider';
+import useConversation from './app/(dashboard)/conversations/useContext';
 
 const ConversationsList: React.FC = () => {
-  const { allConversations, setOpenConversation } =
-    useContext(ConversationContext);
+  const { allConversations, setOpenConversation } = useContext(useConversation);
 
   const updateOpenConvo = async (givenId: number) => {
     setOpenConversation &&
@@ -18,18 +17,22 @@ const ConversationsList: React.FC = () => {
 
   return (
     <>
-      {allConversations.map((conversation, index) => (
-        <div key={`${conversation.id}-${index}`}>
-          <ConversationCard
-            id={conversation.conversation_id}
-            joined_at={conversation.joined_at}
-            conversation_id={conversation.conversation_id}
-            user_id={conversation.user_id}
-            conversations={conversation.conversations}
-            clickHandler={() => updateOpenConvo(conversation.conversation_id)}
-          />
-        </div>
-      ))}
+      {allConversations.length > 0 ? (
+        allConversations.map((conversation, index) => (
+          <div key={`${conversation.id}-${index}`}>
+            <ConversationCard
+              id={conversation.conversation_id}
+              joined_at={conversation.joined_at}
+              conversation_id={conversation.conversation_id}
+              user_id={conversation.user_id}
+              conversations={conversation.conversations}
+              clickHandler={() => updateOpenConvo(conversation.conversation_id)}
+            />
+          </div>
+        ))
+      ) : (
+        <p>There are no active conversations</p>
+      )}
     </>
   );
 };
